requirements.txt
my-iqoptionapi
numpy
schedule
python-telegram-bot
